Made game a little easier.
Health was all the while int. That's why a -0.5f was still causing so much damage. It's not changed to float. Also removed one player spawn location because it was too near the door.

+++ core/src/com/mygdx/amaze/entities/Player.java
@@ -29,7 +29,7 @@ public class Player {
 
     public boolean todestroy;
 
-    public int health = 99;
+    public float health = 99;
     public boolean shielded = false;
     public boolean attacked = false;
     public boolean collidableWithHole = false;
